changing id from int to long in location and relating methods, classes

+++ server/src/main/java/com/codecool/controller/LocationController.java
@@ -30,7 +30,7 @@ public class LocationController {
   }
 
   @PostMapping
-  public int addLocation(@RequestBody NewLocationDTO location) {
+  public long addLocation(@RequestBody NewLocationDTO location) {
     return locationService.addLocation(location);
   }
